pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-deep-coder"
version = "0.1.0"
description = "AI programming assistant built on LangChain Deep Agents"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.5.0",
    "langchain-anthropic>=1.4.0",
    "langchain-openai>=0.3.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai-deep-coder = "coder.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/coder"]
